allow transform to select a subset of columns from the original dataset
