+++ pytorch_pfn_extras/dataset/tabular/tabular_dataset.py
@@ -215,7 +215,7 @@ class TabularDataset(Dataset):
         """
         return ppe.dataset.tabular._join._Join(self, *datasets)
 
-    def transform(self, keys, transform):
+    def transform(self, keys, transform, fetch_keys=None):
         """Apply a transform to each example.
 
         Args:
@@ -224,12 +224,15 @@ class TabularDataset(Dataset):
                 and returns transformed example. :attr:`mode` of
                 transformed dataset is determined by the transformed
                 examples.
+            fetch_keys (tuple of strs): Optional, if defined
+                it will be passed to the dataset to avoid loading
+                uneeded columns
 
         Returns:
             A transfromed dataset.
         """
         return ppe.dataset.tabular._transform._Transform(
-            self, keys, transform)
+            self, keys, transform, fetch_keys)
 
     def transform_batch(self, keys, transform_batch):
         """Apply a transform to examples.
